pyproject.toml
[tool.poetry]
name = "aegisclaim-engine"
version = "0.1.0"
description = "AI-powered medical insurance claim processing engine"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
google-generativeai = "^0.3.2"
PyPDF2 = "^3.0.1"
PyMuPDF = "^1.23.8"
Pillow = "^10.1.0"
numpy = "^1.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytesseract = "^0.3.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.25.1"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.13.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
show_error_context = true
pretty = true
show_column_numbers = true
show_error_end = true

[[tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = true
